6- API_Usuários/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
PyJWT==1.7.1
sqlparse==0.5.3
tzdata==2025.2
